add new types for better json unmarshalling
