+++ storyteller-api/data/defined.go
@@ -1,7 +1,7 @@
 package data
 
 // TODO: Remove this thing after dealing with DB
-var Categories = []Category{
+var CategoriesDefined = Categories{Categories: []Category{
 	{1, "Сеттинг", []*Tag{
 		{1, "Зима", "test", `<svg xmlns="http://www.w3.org/2000/svg" width="512" height="512" viewBox="0 0 512 512">
   <path style="fill:none;stroke:#000;stroke-linecap:round;stroke-linejoin:round;stroke-width:32px" d="M256 32v448m57.72-400A111.47 111.47 0 0 1 256 96a111.47 111.47 0 0 1-57.72-16m0 352a112.11 112.11 0 0 1 115.44 0m136.27-288L62.01 368m375.26-150a112.09 112.09 0 0 1-57.71-100M74.73 294a112.09 112.09 0 0 1 57.71 100M62.01 144l387.98 224M74.73 218a112.09 112.09 0 0 0 57.71-100m304.83 176a112.09 112.09 0 0 0-57.71 100"/>
@@ -47,4 +47,5 @@ var Categories = []Category{
 </svg>
 `},
 	}},
+},
 }
